Fix build and try adding redundancy

+++ src/Software/PTP_Stack/AppServices/PortService.h
@@ -52,7 +52,7 @@ class cPortService : public cBasicService
         // Types
         enum class  PortServiceEvent
         {
-            TIMEOUT = BasicServiceEvent::NEXT_FREE_ENTRY,
+            TIMEOUT = (int)BasicServiceEvent::NEXT_FREE_ENTRY,
             NEXT_FREE_ENTRY,
         };
